Move validation logic from the REST CONTROLLER to the Service layer

+++ src/main/java/com/pascal/api/blogger/service/BlogService.java
@@ -3,9 +3,13 @@ package com.pascal.api.blogger.service;
 import com.pascal.api.blogger.entity.Blog;
 import com.pascal.api.blogger.repository.BlogRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -17,22 +21,62 @@ public class BlogService {
         this.blogRepository = blogRepository;
     }
 
-    public List<Blog> getAllBlogs() {
-        return blogRepository.findAll();
+
+
+    // GET ALL DATA MAIN METHOD
+    // GET ALL DATA WITH TAG VALIDATION
+    public List<Blog> getAllBlogs(String tag) {
+        if (tag != null && !tag.isEmpty()) {
+            return getBlogsByTag(tag);
+        } else {
+            return blogRepository.findAll();
+        }
     }
 
+    // SAVE BLOG MAIN METHOD
+    // WITH VALIDATION CHECK BY RESULT IF HAS ERRORS
+    public ResponseEntity<String> saveBlog(Blog blog ,BindingResult result) {
+        if (result.hasErrors()) {
+            return new ResponseEntity<>(
+                    Objects.requireNonNull(result.getFieldError()).getDefaultMessage(),
+                    HttpStatus.BAD_REQUEST);
+        }
+        blogRepository.save(blog);
+        return new ResponseEntity<>("El blog fue creado exitosamente", HttpStatus.CREATED);
+    }
+
+    // UPDATE BLOG MAIN METHOD
+    // WITH VALIDATION CHECK BY RESULT IF HAS ERRORS
+    // CHECK IF BLOG ID EXISTS
+    public ResponseEntity<String> updateBlog(Blog blog, BindingResult result, Long id) {
+        if (result.hasErrors()) {
+            return new ResponseEntity<>(
+                    Objects.requireNonNull(result.getFieldError()).getDefaultMessage(),
+                    HttpStatus.BAD_REQUEST);
+        }
+        if (getBlogById(id).isEmpty()) {
+            return new ResponseEntity<>("El blog con id: " + id + " no existe", HttpStatus.NOT_FOUND);
+        }
+        blog.setId(id);
+        blogRepository.save(blog);
+        return new ResponseEntity<>("El blog fue modificado exitosamente", HttpStatus.OK);
+    }
+
+    // DELETE BLOG BY ID
+    public ResponseEntity<String> deleteBlogById(Long id) {
+        if (getBlogById(id).isEmpty()) {
+            return new ResponseEntity<>("El blog con id: " + id + " no existe", HttpStatus.NOT_FOUND);
+        }
+        blogRepository.deleteById(id);
+        return new ResponseEntity<>("El blog con id: " + id + " fue eliminado exitosamente", HttpStatus.OK);
+    }
+
+    // GET BLOGS BY ID
     public Optional<Blog> getBlogById(Long id) {
         return blogRepository.findById(id);
     }
 
-    public void saveOrUpdateBlog(Blog blog) {
-        blogRepository.save(blog);
-    }
-
-    public void deleteBlogById(Long id) {
-        blogRepository.deleteById(id);
-    }
-
+    // GET BLOGS BY TAG
     public List<Blog> getBlogsByTag(String tag) {
         return blogRepository.findByTag(tag);
     }
